+++ pulse-binding/CHANGELOG.md
@@ -16,6 +16,8 @@
     - Added `volume::VolumeDB::MINUS_INFINITY` replacement.
     - Deprecated `time::{USEC_INVALID, USEC_MAX}`.
     - Added `time::MicroSeconds::{INVALID, MAX}` replacements.
+    - Deprecated `context::subscribe::{FACILITY_MASK, OPERATION_MASK}`.
+    - Added `context::subscribe::{Facility::MASK, Operation::MASK}` replacements.
  * Fixed failure to update `pa_latest` feature in v2.17 to point to `pa_v14` instead of `pa_v13`.
 
 # 2.19.0 (December 8th, 2020)